+++ tests/Pest.php
@@ -1,5 +1,5 @@
 <?php
 
-use Suleymanozev\FilamentRadioButtonField\Tests\TestCase;
+use Jodeveloper\FilamentRating\Tests\TestCase;
 
 uses(TestCase::class)->in(__DIR__);